chore: Refactor Test Command.
Loading rego files + data files has been refactored into a seperate struct. The Test command now also uses the Runner pattern.
An Engine class has been created to execute queries against the policies.
